+++ lib/Predis/PubSub/AbstractPubSubContext.php
@@ -13,7 +13,6 @@ namespace Predis\PubSub;
 
 use Predis\ClientException;
 use Predis\ClientInterface;
-use Predis\Helpers;
 use Predis\NotSupportedException;
 
 /**